Refactor test automation setup: - Update namespaces for better organization and consistency. - Introduce `[UsedImplicitly]` attribute in shared tests. - Adjust visibility of constants and methods. - Modify method calls to enhance code clarity and maintainability.
