+++ tests/InfiniFrameAutomationTests.WebApp.React/JavascriptInteropTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrameAutomationTests.Tests;
 using InfiniFrameAutomationTests.WebApp.React.TestUtility;
 
 namespace InfiniFrameAutomationTests.WebApp.React;
