+++ tests/InfiniFrameAutomationTests.WebApp.Vue/WebviewWindowTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrameAutomationTests.Tests;
 using InfiniFrameAutomationTests.WebApp.Vue.TestUtility;
 
 namespace InfiniFrameAutomationTests.WebApp.Vue;
