+++ tests/InfiniFrameAutomationTests/AutomationPortUtility.cs
@@ -1,8 +1,13 @@
+// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
 using System.Net;
 using System.Net.Sockets;
 
 namespace InfiniFrameAutomationTests;
-
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
 public static class AutomationPortUtility {
     public static int GetAvailablePort() {
         using var listener = new TcpListener(IPAddress.Loopback, 0);
