+++ tests/InfiniFrameAutomationTests/IAutomationPage.cs
@@ -1,5 +1,5 @@
 // ---------------------------------------------------------------------------------------------------------------------
-// Code
+// Imports
 // ---------------------------------------------------------------------------------------------------------------------
 namespace InfiniFrameAutomationTests;
 // ---------------------------------------------------------------------------------------------------------------------
