+++ tests/InfiniFrameAutomationTests/Selenium/SeleniumAutomationPage.cs
@@ -1,10 +1,10 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-using OpenQA.Selenium;
 using System.Text.Json;
+using OpenQA.Selenium;
 
-namespace InfiniFrameAutomationTests;
+namespace InfiniFrameAutomationTests.Selenium;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
