+++ tests/InfiniFrameAutomationTests/Tests/SharedJavascriptInteropTests.cs
@@ -3,8 +3,9 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniFrame;
 using InfiniFrameTests.Shared;
+using JetBrains.Annotations;
 
-namespace InfiniFrameAutomationTests;
+namespace InfiniFrameAutomationTests.Tests;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
@@ -13,10 +14,14 @@ public abstract class SharedJavascriptInteropTests : InfiniFrameAutomationTestBa
 
     protected abstract string TitleToggleButtonSelector { get; }
 
-    protected virtual string ToggledTitle => "New Title";
+    private const string ToggledTitle = "New Title";
 
+    // -----------------------------------------------------------------------------------------------------------------
+    // Methods
+    // -----------------------------------------------------------------------------------------------------------------
     [Test]
     [NotInParallel(ParallelControl.Playwright)]
+    [UsedImplicitly]
     public async Task FullscreenHtmlButton_ShouldToggleInfiniFrameFullscreen() {
         bool originalFullscreenState = RuntimeContext.Window.FullScreen;
         IAutomationPage page = await GetRootPageAsync();
@@ -40,6 +45,7 @@ public abstract class SharedJavascriptInteropTests : InfiniFrameAutomationTestBa
 
     [Test]
     [NotInParallel(ParallelControl.Playwright)]
+    [UsedImplicitly]
     public async Task TitleHtmlButton_ShouldToggleInfiniFrameTitle() {
         IAutomationPage page = await GetRootPageAsync();
         string originalTitleState = RuntimeContext.Window.Title;
