+++ tests/InfiniFrameAutomationTests/Tests/SharedWebviewWindowTests.cs
@@ -1,15 +1,18 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-using InfiniFrameTests.Shared;
 
-namespace InfiniFrameAutomationTests;
+using InfiniFrameTests.Shared;
+using JetBrains.Annotations;
+
+namespace InfiniFrameAutomationTests.Tests;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
 public abstract class SharedWebviewWindowTests : InfiniFrameAutomationTestBase {
     [Test]
     [NotInParallel(ParallelControl.Playwright)]
+    [UsedImplicitly]
     public async Task Title_ShouldBeExpectedValue() {
         IAutomationPage page = await GetRootPageAsync();
